Tenants feed is now filtered by date created then priority
