+++ src/app/services/pipes/issues/filter-issues.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from "@angular/core";
 import { FilterIssuesByHouse, FilterIssuesByStatus } from "./filter-issues.pipe";
-import { OrderByDate } from "./order-by-date.pipe";
+import { OrderByDate, OrderByDatePriority } from "./order-by.pipe";
 
 @NgModule({
   imports: [],
-  declarations: [ FilterIssuesByHouse, FilterIssuesByStatus, OrderByDate ],
-  exports: [ FilterIssuesByHouse, FilterIssuesByStatus, OrderByDate ],
+  declarations: [ FilterIssuesByHouse, FilterIssuesByStatus, OrderByDate, OrderByDatePriority],
+  exports: [ FilterIssuesByHouse, FilterIssuesByStatus, OrderByDate, OrderByDatePriority ],
 })
 export class IssueFilters { }
