--- src/app/services/pipes/issues/filter-issues.pipe.ts.orig
@@ -1,27 +1,0 @@
-import { Pipe, PipeTransform } from "@angular/core";
-import IIssueResponse from "src/app/models/response/landlord/issue-response.interface";
-import IHouseResponseLandlord from "src/app/models/response/landlord/house-response.interface";
-
-<<<<<<< HEAD:src/app/services/pipes/issues/filter-issues.pipe.ts
-@Pipe({ name: "FilterIssuesByHouse" }) 
-export class FilterIssuesByHouse implements PipeTransform {
-  transform(issues: IIssueResponse[], currentHouse: IHouseResponse) {
-    if (currentHouse.id == null) {
-=======
-@Pipe({ name: "FilterIssues" }) 
-export class FilterIssuesPipe implements PipeTransform {
-  transform(issues: IIssueResponse[], currentHouse: IHouseResponseLandlord) {
-    if (currentHouse == null) {
->>>>>>> develop:src/app/services/pipes/filter-issues.pipe.ts
-      return issues;
-    }
-    return issues.filter((element, index, array) => element.house.id === currentHouse.id);
-  }
-}
-
-@Pipe({ name: "FilterIssuesByStatus" }) 
-export class FilterIssuesByStatus implements PipeTransform {
-  transform(issues: IIssueResponse[], isResolved: boolean) {
-    return issues.filter((issue, index, array) => issue.isResolved === isResolved);
-  }
-}
