+++ src/app/views/landlord/issue-feed/issue-feed.page.html
@@ -22,7 +22,7 @@
       </ion-refresher-content>
     </ion-refresher>
   <div *ngIf="issues" class="issues-list">
-    <app-issue-card *ngFor="let issue of (issues | FilterIssuesByHouse:currentHouse | FilterIssuesByStatus:viewingResolved)" [issue]="issue" (click)="viewIssue(issue)"></app-issue-card>
+    <app-issue-card *ngFor="let issue of (issues | FilterIssuesByHouse:currentHouse | FilterIssuesByStatus:viewingResolved | OrderByDatePriority)" [issue]="issue" (click)="viewIssue(issue)"></app-issue-card>
   </div>
 
   <div *ngIf="!issues">
